fix: fixed bug when fetching devices
