+++ src/store/devices/index.tsx
@@ -8,14 +8,13 @@ export type Device = {
     name: string
     room: string
     amount?: number
+    user_id : string
     [key : string]: any
 }
 
 type ActionProps = {
-    addDevice : (device: Device) => void
-    removeDevice : (device: Device) => void,
-    alterDevice : (device: Device, prop: string, value: string | number) => void
-    addDevices : (devices: Device[]) => void
+    addDevices : (devices: Device[]) => void,
+    clearDeviceStore : () => void
 }
 
 type StoreProps = {
@@ -23,25 +22,10 @@ type StoreProps = {
     actions: ActionProps
 }
 
-export const useDeviceStore = create<StoreProps>((set, get) => ({
+export const useDeviceStore = create<StoreProps>((set) => ({
     devices : [],
     actions : {
-        addDevice : (newDevice) => {
-            set((state) => ({devices : [...state.devices, newDevice]}))
-        },
-        removeDevice : (deviceToRemove) => set((state) => ({
-            devices : state.devices.filter((device) => device.id !== deviceToRemove.id)
-        })),
-        alterDevice : (device, prop, value) => {
-            const deviceToAlter = get().devices.find(item => device === item)
-            if(deviceToAlter) {
-                deviceToAlter[prop] = value
-                console.log(get().devices)
-                // const newState = get().devices.filter((theDevices) => theDevices.id !== deviceToAlter.id) 
-                // set((state) => ({devices : [...newState, deviceToAlter]}))
-            }
-            
-        },
-        addDevices : (devices) => set(({devices : devices}))
+        addDevices : (devices) => set(({devices : devices})),
+        clearDeviceStore : () => set(() => ({devices : []}))
     }
 }))
